[IDEA References] Return `null` from `org.jetbrains.kotlin.idea.references.KDocReference.resolve` in case of multiple found symbols
This change is required to be able to show a drop-down list in IDEA when trying to perform "Go To Declaration" on KDoc references with multiple resolved results.
In cases when `resolve` returns `null`, IDE will use `org.jetbrains.kotlin.idea.references.AbstractKtReference.multiResolve` instead and show all the found results.

^KT-76607