[Analysis API] Adjust `KDocReferenceResolver` to return a set of only non-shadowed symbols
It's vital that the resolver returns a set of only visible symbols, because the returned result is also used for "Find usages", import optimizations and other actions in the IDE.
So if there are any unrelated symbols, then they directly affect the user experience.
That's why the resolver should only return the symbol the KDoc name is pointing to (i.e. the first found symbol w.r.t priorities) or a set of referred names (if there are any overloads).

^KT-76607