[JVM] Support evaluation of local function refs capturing local vars
To achieve that, we share `JvmLocalDeclarationsLowering` data between
files in case of code fragment compilation since code fragment may
refer a local declaration from another file.

After such generalization of this lowering and implementation the logic
of adding parameters to debugger-generated entry evaluation entry point,
we do not need `FragmentLocalFunctionPatchLowering` as a separate
lowering anymore.

Also, since `JvmLocalDeclarationsLowering` relies on
`JvmEvaluatorData#evaluatorGeneratedFunction` from now on, we have to
initialize it in K1 setup as well.

^KT-73013: Fixed